// 4-inventory-dashboard/src/app/shared/Enums/EquipmentEnums.ts
export enum EquipmentCategory
{
    Tools = 1,
    Electronics = 2,
    Vehicles = 3,
    Furniture = 4,
    SafetyGear = 5,
    Other = 6
}

export enum EquipmentCondition
{
    New = 1,
    Good = 2,
    Damaged = 3,
    Retired = 4
}

export enum EquipmentStatus
{
    Available = 1,
    InUse = 2,
    Reserved = 3,
    UnderMaintenance = 4
}